fix: hack about joining room
For some reason, when using Dendrite after an update at one point
If trying to join the room while already joined, dendrite will complain
that you're not in the room, even though you are. So, remove the join
room part. This will fuck things up on the first run b/c then a join is
needed. Will sort it out later